pyproject.toml
[tool.poetry]
name = "booking-site"
version = "0.1.0"
description = "Booking Site"
repository = "https://www.github.com/Zontax/booking-site"
readme = "README.md"
authors = ["Zontax <jrvadim04@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.11"
pytest = "8.2.2"
gunicorn = "^22.0.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.4.2"
django-simple-captcha = "^0.6.0"
django-recaptcha = "^4.0.0"
django-phonenumber-field = "^7.3.0"
django-colorfield = "^0.11.0"
humanize = "^4.9.0"
mimesis = "^17.0.0"
bleach = "^6.1.0"
celery = "^5.4.0"
pytils = "^0.4.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
termcolor = "^2.4.0"
beautifulsoup4 = "^4.12.3"
django-filebrowser = "^4.0.3"
django-compressor = "^4.5"
django-bootstrap5 = "^24.2"
drf-spectacular = "^0.27.2"
libsass = "^0.23.0"
django-sass-processor = "^1.4.1"
pyyaml = "^6.0.1"
ua-parser = "^0.18.0"
user-agents = "^2.2.0"
redis = "^5.0.6"
django-admin-extra-buttons = "^1.5.8"
selenium = "^4.22.0"
django-extensions = "^3.2.3"
phonenumbers = "^8.13.39"
werkzeug = "^3.0.3"
pygments = "^2.18.0"
httpie = "^3.2.2"
django-allauth = "^0.63.3"
flower = "^2.0.1"
email-validator = "^2.2.0"
django-import-export = "^4.1.1"
django-admin-interface = "^0.28.8"
django-json-widget = "^2.0.1"
django-image-uploader-widget = "^0.5.0"
django-tinymce = "^4.1.0"
django-adminactions = "^2.3.0"
reportlab = "^4.2.2"
django-silk = "^5.1.0"
[tool.poetry.dev-dependencies]
pytest = "8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
somecommand = "src.main:run"
